Fix attached image upload
I had to switch the 'directory' page creation so it will only happen if there's actual Markdown content in it. The reason is that we have images and Markdown in the doc directory now, and some dirs may only be for images. I also had to make the error about non-Markdown files into a warning, though I did that earlier.

The fix here required changing the confluence storage format markup injection from the python code over to the lua code, and then using the image-scanning method in the python code to inject a special 'protocol' called **confluence-attachment:** which we detect in the lua and replace it with a reference to an attached image.