Refuse to overwrite pages with mismatched parents
You run into this problem if you have multiple instances of this
script writing to the same confluence namespace.  One git repo may
want to write the page "FooBar" with a parent page for one org and
another git repo may want to write the page "FooBar" with a parent
page for a different org -- all in the same namespace.

Before this change, markdown-to-confluence would happily flip flop
that content back and forth.

With this change, the first run will succeed and claim the page
"FooBar", setting a particular parent page.  The second run will
find the page "FooBar", find that the existing parent page in
confluence does not match its requested parent page - and fail
loudly.